% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_setup.R
\name{cloud_setup}
\alias{cloud_setup}
\title{Create DigitalOcean droplet setup to deploy shiny apps}
\usage{
cloud_setup(droplet, ...)
}
\arguments{
\item{droplet}{droplet object created with analogsea package, if creating a new droplet, leave empty}

\item{...}{additional params to function analogsea::droplet_create}
}
\description{
Create DigitalOcean droplet setup to deploy shiny apps
}
