% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_setup.R
\name{cloud_setup}
\alias{cloud_setup}
\title{Create DigitalOcean droplet setup to deploy \code{Shiny} apps}
\usage{
cloud_setup(size = 1, ...)
}
\arguments{
\item{size}{initial RAM (in GBs) for the droplet. Available values: 1, 2, 4, 8. You can always resize the droplet later on.}

\item{...}{additional params to function analogsea::docklet_create}
}
\description{
This function creates a new DigitalOcean droplet with Docker 19.03.12 on Ubuntu 18.04 with every required package and service to run xai2shiny applications in the cloud.
}
